% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjbae.R
\name{cjbaemce}
\alias{cjbaemce}
\alias{cjbae_plot}
\title{Bayesian inference for conjoint experiments}
\usage{
cjbaemce(data, formula, id)

cjbae_plot(posteriors, visuals = NULL)
}
\description{
Bayesian inference for conjoint experiments
}
