% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjbae_formula.R
\name{cjbae_formula}
\alias{cjbae_formula}
\title{Conjoint formula conversion for Bayesian analysis}
\usage{
cjbae_formula(data, formula, id)
}
\arguments{
\item{data}{A tidy, long-format conjoint dataframe.}

\item{formula}{A formula specifying an AMCE/MM model, to be converted for bayesian analysis. All variables should be factors.}

\item{id}{A variable, within data, containing respondent IDs.}
}
\description{
Convert standard tidy conjoint function (data, formula, id) input into the input for a `brms` model.
}
\details{
\code{cjbae_formula} takes three arguments: `data`, `formula`, and `id`, and uses these to specify a formula to be passed to `brm()`, to facilitate bayesian analysis of conjoint experiments otherwise analysed using `cregg`. This function is integrated into other `cjbae()` functions and only necessary in isolation if you are directly using `brms()`.
}
\examples{
#load example data from cregg
library(cregg)
data(immigration)
#with simple input formula
cjbae_formula(data = immigration, formula = ChosenImmigrant ~ Gender + Job, id = CaseID)
#or matched positionally
cjbae_formula(immigration, ChosenImmigrant ~ Gender + Job, CaseID)
#can pre-specify longer input formula
f1 <- ChosenImmigrant ~ Gender + Education + LanguageSkills +
CountryOfOrigin + Job + JobExperience + JobPlans +
ReasonForApplication + PriorEntry
cjbae_formula(immigration, f1, CaseID)
}
