% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjbae_plot.R
\name{cjbae_plot}
\alias{cjbae_plot}
\title{Simple plotting of cjbae outputs.}
\usage{
cjbae_plot(data, visual = c("ridge", "halfeye"), estimate = c("amce",
  "mm"))
}
\arguments{
\item{data}{A tidy dataframe of either AMCEs or MMs.}

\item{visual}{Currently either "ridge" or "halfeye" - two different takes on a distribution plot.}

\item{estimate}{Either "mm" or "amce".}
}
\value{
A plot of parameter distributions.
}
\description{
Visualise Bayesian marginal means and AMCEs
}
\details{
\code{cjbae_plot()} plots AMCEs or marginal means of feature-levels as distributions, colour-coded by feature.
}
\examples{
#' #load example dataset from {cregg} (Leeper 2019)
library(cregg)
data(taxes)
# formula
f <- chose_plan ~ taxrate1 + taxrate2 + taxrate3 + taxrate4 + taxrate5 + taxrate6 + taxrev
# prior - minimally informative
prior <- c(set_prior("normal(0, .2)", class = "Intercept"),
set_prior("normal(0, .2)", class = "b"),
set_prior("exponential(10)", class = "sd"),
set_prior("exponential(10)", class = "sigma"))
# run amce function with save specified, saves brmsfit to working directory - this will take a while
amce_bae(data = taxes, formula = f, id = ID, prior = prior, save_amce = TRUE)
# run mm function on the saved output
readRDS(baerms)
mm <- mm_bae(baerms, f, ID)
# plot MMs
cjbae_plot(mm, "ridge", "mm")
# generate df of AMCEs
amce <- cjbae_df(baerms)
# plot AMCEs
cjbae_plot(amce, "halfeye", "amce")
}
